+++ DogCare/Pages/Schedule.xaml
@@ -3,7 +3,8 @@
     xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:local="clr-namespace:DogCare;assembly=DogCare"
-             x:Class="DogCare.Schedule">
+             x:Class="DogCare.Schedule"
+             >
     <ContentView>
         <Grid x:Name="grid_layout">
             <Grid.RowDefinitions>
@@ -16,9 +17,10 @@
                     <ColumnDefinition Width="0.6*" />
                     <ColumnDefinition Width="0.2*" />
                 </Grid.ColumnDefinitions>
+                
                 <Button x:Name="scheduleView_Button" Grid.Column="0"  BackgroundColor="#D3D3D3" HorizontalOptions= "Start" VerticalOptions="Center"  />
-                <Label x:Name="Month_Button" Grid.Column="1" VerticalOptions="Center" FontSize="20" TextColor="Black" BackgroundColor="#D3D3D3" ></Label>
-                <Button x:Name="Editor_Button"  Grid.Column="2" BackgroundColor="#D3D3D3" VerticalOptions="Center"  HorizontalOptions="End" />
+                <Label x:Name="Month_Button" Text="Month" Grid.Column="1" VerticalOptions="Center" FontSize="20" TextColor="Black" BackgroundColor="#D3D3D3" ></Label>
+                <Button x:Name="Editor_Button" Text="Month" Grid.Column="2" BackgroundColor="#D3D3D3" VerticalOptions="Center"  HorizontalOptions="End" />
             </Grid>
 
             <ListView x:Name="scheduleview_list" HasUnevenRows="False"
