--- DogCare/Schedule.xaml
@@ -1,35 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns:schedule="clr-namespace:Syncfusion.SfSchedule.XForms;assembly=Syncfusion.SfSchedule.XForms"
-    xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:DogCare;assembly=DogCare"
-             x:Class="DogCare.Schedule"
-             >
-    <ContentView>
-        <Grid x:Name="grid_layout">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="0.1*" />
-                <RowDefinition Height="0.9*" />
-            </Grid.RowDefinitions>
-            <Grid Grid.Row="0" x:Name="header_layout" BackgroundColor="#D3D3D3" >
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="0.2*" />
-                    <ColumnDefinition Width="0.6*" />
-                    <ColumnDefinition Width="0.2*" />
-                </Grid.ColumnDefinitions>
-                <Button x:Name="monthView" Grid.Column="0" Text="month"/>
-                <Button x:Name="scheduleView_Button" Grid.Column="1"  BackgroundColor="#D3D3D3" HorizontalOptions= "Start" VerticalOptions="Center"  />
-                <Label x:Name="Month_Button" Grid.Column="2" VerticalOptions="Center" FontSize="20" TextColor="Black" BackgroundColor="#D3D3D3" ></Label>
-                <Button x:Name="Editor_Button"  Grid.Column="3" BackgroundColor="#D3D3D3" VerticalOptions="Center"  HorizontalOptions="End" />
-            </Grid>
-
-            <ListView x:Name="scheduleview_list" HasUnevenRows="False" 
-					HeightRequest="200" RowHeight="50" WidthRequest="200" 
-					VerticalOptions="Start" HorizontalOptions="Start" 
-					SeparatorVisibility="Default" SeparatorColor="Gray" Grid.Row="1"
-				BackgroundColor="White">
-            </ListView>
-            
-        </Grid>
-    </ContentView>
-</ContentPage>
